+++ src/table/utils/constants.ts
@@ -13,7 +13,7 @@ export enum TABLE_STATUSCODE {
 }
 
 export const DEFAULT_TABLE_CONTEXT_PATH = "azurite_table_context";
-export const TABLE_API_VERSION = "2020-06-12";
+export const TABLE_API_VERSION = "2020-08-04";
 export const VERSION = "3.12.0";
 
 export const HeaderConstants = {