+++ ChangeLog.md
@@ -7,6 +7,7 @@
 General:
 
 - updated typescript to 4.2.4
+- Bump up Azure Storage service API version to 2020-08-04.
 
 Blob: