+++ ChangeLog.md
@@ -18,6 +18,7 @@ Table:
 - added tests for invalid etag format
 - corrected code to support typescript 4.2.4 update
 - added check for invalid table name
+- improve handling of empty strings and strings with multiple spaces for query filters
 
 ## 2021.4 Version 3.12.0